+++ hack/upgrade.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 
-rm /download/install.sh
+if [ -f /download/install.sh ]; then
+	rm /download/install.sh
+fi
+
 if [ -f /data/etc/upgradeurl ]; then
 	DL_URL=`head -n 1 /data/etc/upgradeurl`
 	echo "Downloading from " $DL_URL >> $BASEDIR/install.log
@@ -8,5 +11,6 @@ if [ -f /data/etc/upgradeurl ]; then
 else
 	DL_URL=https://raw.github.com/boxeehacks/boxeehack/master/install/install.sh
 fi
+
 curl $DL_URL -o /download/install.sh
 sh /download/install.sh &
